+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.0.6-alpha.0
+- Add number type shapes, Total count 115
+- Change the index to start from 0
 ## 0.0.5-alpha.1
 - Add more shapes. Total count 105
 - Fix noise toggle